:sparkles: Merch page set up with product grid component and fake products listed
