+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "catnadianparty",
-  "version": "0.0.9",
+  "version": "0.0.10",
   "homepage": "https://leondoes.github.io/catnadian-party/",
   "private": true,
   "dependencies": {
