+++ src/assets/products.js
