+++ src/components/ProductGrid/index.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+const ProductGrid = ({ products }) => {
+  return (
+    <div className="product-grid">
+      {products.map(product => (
+        <div key={product.id} className="product">
+          <img src={product.image} alt={product.name} />
+          <h3>{product.name}</h3>
+          <p>{product.price}</p>
+        </div>
+      ))}
+    </div>
+  );
+};
+
+export default ProductGrid;
