+++ src/pages/MerchandisePage/index.js
@@ -1,21 +1,13 @@
-/*
-Product Grid Component: Displaying party-themed merchandise like t-shirts, mugs, etc.
-*/
-
 import React from 'react';
+import ProductGrid from '../../components/ProductGrid';
+import products from '../../assets/products';
 
-const ProductGrid = ({ products }) => {
+const MerchandisePage = () => {
   return (
-    <div className="product-grid">
-      {products.map(product => (
-        <div key={product.id} className="product">
-          <img src={product.image} alt={product.name} />
-          <h3>{product.name}</h3>
-          <p>{product.price}</p>
-        </div>
-      ))}
+    <div>
+      <ProductGrid products={products} />
     </div>
   );
-};
+}
 
-export default ProductGrid;
+export default MerchandisePage;
